fix(android-emulator): resolve remaining golangci-lint findings
- unzip.go: errorlint flagged comparing io.EOF with != instead of errors.Is.
- diskspace_unix.go: unconvert flagged uint64(stat.Bavail) as redundant --
  Bavail is already uint64 on both linux and darwin, verified by
  cross-compiling for both before removing the conversion.

+++ backend/internal/androidsdk/diskspace_unix.go
@@ -14,5 +14,5 @@ func FreeBytes(path string) (uint64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, fmt.Errorf("androidsdk: statfs %s: %w", path, err)
 	}
-	return uint64(stat.Bavail) * uint64(stat.Bsize), nil //nolint:gosec // G115: block size/count from the kernel's own statfs result, never negative in practice
+	return stat.Bavail * uint64(stat.Bsize), nil //nolint:gosec // G115: block size/count from the kernel's own statfs result, never negative in practice
 }
